backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mysql-connector-python==8.2.0
pydantic==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0
